+++ Taller/NewtonNu.h
@@ -1,6 +1,6 @@
 #ifndef NEWTONNU_H_INCLUDED
 #define NEWTONNU_H_INCLUDED
 
-void newtonnu(double, double, double salida[2]);
+void newtonnu(long double, long double, long double salida[2]);
 
 #endif // NEWTONNU_H_INCLUDED
