+++ Taller/PoleMatrix.h
@@ -1,5 +1,5 @@
 #ifndef POLEMATRIX_H_INCLUDED
 #define POLEMATRIX_H_INCLUDED
 
-void  PoleMatrix (double xp, double yp, double matrizTransformada[3][3]);
+void  PoleMatrix (long double xp, long double yp, long double matrizTransformada[3][3]);
 #endif // POLEMATRIX_H_INCLUDED
